+++ useServerContext.test.mjs
@@ -1,11 +1,12 @@
 // @ts-check
 
-import { cleanup, renderHook } from "@testing-library/react-hooks/lib/pure.js";
 import { strictEqual } from "assert";
 import React from "react";
+import ReactTestRenderer from "react-test-renderer";
 
 import ServerContextContext from "./ServerContextContext.mjs";
 import assertBundleSize from "./test/assertBundleSize.mjs";
+import ReactHookTest from "./test/ReactHookTest.mjs";
 import useServerContext from "./useServerContext.mjs";
 
 /**
@@ -21,50 +22,69 @@ export default (tests) => {
   });
 
   tests.add("`useServerContext` with server context context missing.", () => {
-    try {
-      const { result } = renderHook(() => useServerContext());
+    /** @type {Array<unknown>} */
+    const results = [];
 
-      strictEqual(result.all.length, 1);
-      strictEqual(result.current, undefined);
-      strictEqual(result.error, undefined);
-    } finally {
-      cleanup();
-    }
+    ReactTestRenderer.act(() => {
+      ReactTestRenderer.create(
+        React.createElement(ReactHookTest, { hook: useServerContext, results })
+      );
+    });
+
+    strictEqual(results.length, 1);
+    strictEqual(results[0], undefined);
   });
 
   tests.add("`useServerContext` getting the server context.", () => {
-    try {
-      const serverContextA =
-        /** @type {import("./ServerContextContext.mjs").ServerContext} */ ({});
-      const { result, rerender } = renderHook(() => useServerContext(), {
-        wrapper: ({ serverContext, children }) =>
-          React.createElement(
-            ServerContextContext.Provider,
-            { value: serverContext },
-            children
-          ),
-        initialProps: {
-          serverContext: serverContextA,
-        },
-      });
+    /** @type {Array<unknown>} */
+    const results = [];
 
-      strictEqual(result.all.length, 1);
-      strictEqual(result.current, serverContextA);
-      strictEqual(result.error, undefined);
+    /** @type {import("react-test-renderer").ReactTestRenderer | undefined} */
+    let testRenderer;
 
-      // The server context shouldn’t change, but it’s good to test the normal
-      // React context and hooks behavior anyway.
+    const serverContextA =
+      /** @type {import("./ServerContextContext.mjs").ServerContext} */ ({});
 
-      const serverContextB =
-        /** @type {import("./ServerContextContext.mjs").ServerContext} */ ({});
+    ReactTestRenderer.act(() => {
+      testRenderer = ReactTestRenderer.create(
+        React.createElement(
+          ServerContextContext.Provider,
+          { value: serverContextA },
+          React.createElement(ReactHookTest, {
+            hook: useServerContext,
+            results,
+          })
+        )
+      );
+    });
 
-      rerender({ serverContext: serverContextB });
+    strictEqual(results.length, 1);
+    strictEqual(results[0], serverContextA);
 
-      strictEqual(result.all.length, 2);
-      strictEqual(result.current, serverContextB);
-      strictEqual(result.error, undefined);
-    } finally {
-      cleanup();
-    }
+    const tr =
+      /** @type {import("react-test-renderer").ReactTestRenderer} */
+      (testRenderer);
+
+    // The server context shouldn’t change, but it’s good to test the normal
+    // React context and hooks behavior anyway.
+
+    const serverContextB =
+      /** @type {import("./ServerContextContext.mjs").ServerContext} */ ({});
+
+    ReactTestRenderer.act(() => {
+      tr.update(
+        React.createElement(
+          ServerContextContext.Provider,
+          { value: serverContextB },
+          React.createElement(ReactHookTest, {
+            hook: useServerContext,
+            results,
+          })
+        )
+      );
+    });
+
+    strictEqual(results.length, 2);
+    strictEqual(results[1], serverContextB);
   });
 };